requirements.txt
setuptools>=40.3.0
urllib3>=1.25.3
amanobot>=1.3.0
bs4==0.0.1
colorama==0.4.1
youtube-dl>=2019.6.8
duckpy>=2.0.0
python-aiml==0.9.1
schedule==0.6.0
telepot
emoji~=0.5.4
utils~=1.0.1
aiohttp~=3.6.2
keyboard~=0.13.5
Pillow~=6.2.2
beautifulsoup4~=4.8.2
Pyrogram
bitlyshortener
SpeechRecognition
pydub
pymysql
google-api-python-client
ffprobe-python
dropbox
wikipedia
python-dateutil
youtube-search-python
pyspellchecker
numpy==1.14.5
opencv-python
cvlib
tensorflow
telegraph
pyimgur
pypng
pyqrcode